api/requirements.txt
azure-functions
azure-cosmos